Move parse_csp_list_from_metadata out of GlobalScope
